blah blah.

+++ src/Public/stdbool.h
@@ -1,4 +1,4 @@
 #define bool BOOL_
-typedef int BOOL_;
+typedef uint8_t BOOL_;
 #define true 1
 #define false 0
